iQIcBAABCAAGBQJVrwWjAAoJELcr6pgkX/+TlJsP/3/kTil0jlW4TrMNjyU8MqgW  ILU8Ir2qm+XdTiUo8FhHM4yyf6HSeN2sFMVKbSLuFJBV5A1DaESvHz0Q8HD7tBzv  NL3OxKVtJ5mv6jHTjJDfyXZwuTyYRBUk2siiCR3/JaawpBCNG3dsDOJN6qdrcIid  WlDZKKM1YC7IBMthYHTav22QYJ+S0yGtJ/o4W1gqLbvaCP3aR+Mop6u6LeIv5g05  HZ6FTu0AJhPC9M1WAq53iQUWecPnr4Cl6D6VKaakirtBBls387P0GxAmXdKrakS+  c0ouOayAs9SObDG4ZlVytYegEgKI2RZkqoRzXHWEKPdGPjmNe5nc2G+BccL76DSS  YIY5Br36yqR2xz0JlhgEpMewGrgRqpPXy9UMHZ18jDYAqswh+ygqrAM8uMLcIu6R  uDkxwcrU7Ez8kiYBb+IqJw6e37uQcU07I+lWpZJzExMU/LejB4ngwNYl/O3/k4fv  QgTDu6HfEGpQEpOLd4p3Kq1cVFpbokOWKNvfjC74uUsDfORVGC0VvcwnHEzDZDed  hegDlx639hvkKR8shO5E0crncQ0dlecs0cR6nRJk8T2JrtIluy3MbDlLsNQHDA/4  iR3tIo148k9Q8xsTZP4vUbPeVfLz8G5IFX8OhzfBa468lQEaZZgDye9t0fAPs68t  JX0C+s2rlKThzA8zN70X  =XsVy  -----END PGP SIGNATURE-----
 iQIcBAABCAAGBQJVrv8NAAoJELcr6pgkX/+Te4wP/0+IoVLEqk2K0uA20oFpmNqV
 /XYGSoy3PC9SxDBVzeSQD6AKBLsllxBlRJjKSnuFvzjwSUXTFVFLOT51ocopFz2i
 jxeOnI0sw4RelzA5sR7sm/hnNckhybPvokWRWQ1x7bxhjM1cOOYQJgBrt3Kd5082
 ZVqcz6RZqhEq5WpaPP8rB2b417njSVsZncnBkGqh4px3foG4SmBt0j1G1Jxj7Ad2
 So8q8JIc/MGLpu5yqF2EChy3k4wB2+6e9rbg8nVHUZBrSJJ9n6IyqSJl5mAQkNWk
 vM3tj4d6XU1ht13dck64fzNNPsvvY0gRs4iovgSY7YiqyG89Wgq2eT3MtNR/Flcf
 cRN1Ju2nEZRWXkePl8eL5oZOudCUpZOK/0eDD/BA9lVVifLOOHKezmiJfXV1zCDT
 t/tF1vSD1Y26BwuFZl1sSAAd+p1H76ALTfcMPg7MUfLrz7AIbRBicNgsQ5jO3HDQ
 nYehVfgFiv2jHZVew7vKR9umPOIRZ79TDzvbrYxrtmj8xBVrjkgnHutwpb9+I7nX
 yRjuKPtn9W7I+mUGKKrN16FCnzfFcU5uLMXpi4Fp1AYqBq3nBTzNOPAdsBIf1SuK
 YkDHBpAsSbnmmmf95CB9ApCeON6hNuWXuaD9Jj8X4njHCCr4oQ8ClFof172jOyOo
 +Ow8H9n1zKyDgWNNdzRH
 =EGj+
 -----END PGP SIGNATURE-----

 iQEcBAABAgAGBQJVQzd4AAoJEJGWkypDEpFJWN4IAJxJkyaZyStFOY2S72Z7ufk0
 Ay/poKBwXOvly1mXRQbYxcA6GoYEneMmbxTgVWz/u4HZN+VqQOWn+zevhNxpUOif
 IoiCSTDjMcE6wLUe732yEqLcoXapwJINg1HcuHTOaDT70B1HTnS7smUNOIAMp5Kx
 /j2ivvJBX5BYhGLOFCCoV8yTF4nVnu9f2dnRlRtB7ebV77iyc/cd+VyjMYPTbZ28
 cCVkqTKanLKh8qd5v96i9vztomkRd3fUyaT9/Wul4LwK7UZu2tLfJRqYETyGQF0f
 x0K+Zrr7MuE+qq7OdUEFnN/C0ITKvTb76xvpn4FUkSb6bxGmdv0AuOZJRSUb9AQ=
 =rvIW
 -----END PGP SIGNATURE-----

Force checking of ports in appliance policer

Funny thing, if you don't erase the evidence, the Appliance Police
will keep on arresting you. :( This fix forces the Appliance Police to
make a new check of the port availability everytime. We hadn't seen this
issue crop up before because most failures had been happening due to the
UI worker dying. This meant that 443 was still contactable and hence the
cache that the utils.net keeps, did not affect the Police.

If a port is uncontactable, because of network issues for example, then
the fact the port is unreachable was cached in the utils.net module and
because the force flag was absent was never rechecked and the cached
result just served up again and again. Thus continuing slaves just led
to them getting pulled over by the Appliance Police again for crimes
they hadn't committed.